Use single S3 client instance per bucket
Added a new singleton map which stores `AmazonS3Client`s already
created for each S3 bucket so that the clients can be reused. This
avoids creating new connection per each downloaded object, which is
extremely inefficient when thousands of small files are fetched from
the same bucket.

The original sbt plugin does this as well:

https://github.com/frugalmechanic/fm-sbt-s3-resolver/blob/396ea80602b572ec44604b8a77dcd3767d30e421/src/main/scala/fm/sbt/S3URLHandler.scala#L248